Changes to get the server running with npm start
