+++ .eslintrc.json
@@ -1,3 +1,6 @@
 {
-    "extends": "airbnb-base"
+    "extends": "airbnb-base",
+    "rules": {
+        "linebreak-style": "off"
+    }
 }
