+++ .vscode/settings.json
@@ -1,4 +1,3 @@
 {
-    "editor.tabSize": 2,
-    "eslint.autoFixOnSave": true
+    "editor.tabSize": 2
 }
