+++ package.json
@@ -4,7 +4,7 @@
   "description": "Bring Untappd info to Slack",
   "main": "index.js",
   "scripts": {
-    "start": "node index.js",
+    "start": "babel-node index.js",
     "test": "mocha --require babel-polyfill --require babel-register --bail --recursive --sort --full-trace ./test"
   },
   "dependencies": {
